backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pdfplumber
pytesseract
python-dotenv
langchain
Pillow
sentence-transformers
transformers
torch
